Add Docker integration to GitHub Actions
Added a few lines in the GitHub Actions workflow file (dotnet.yml) to build a Docker image of the Wowshop project. The changes now allow the creation of a Docker image and authentication into Docker Hub using secrets. This Docker image is then tagged and pushed to Docker Hub. This integration enables easy, secure distribution of the application in a containerized format, improving deployment efficiency.